requirements.txt
black
ipython
jira
jiracli
keyring
pyyaml
